+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## 0.3.2 - 28-04-22
 
 ### Fixed
 * Override val not respected. [#38](https://github.com/nojaf/telplin/issues/38)